# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ghostwatch API — AI-powered sports betting assistant
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"
  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "204":
          description: Deleted
  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a message and receive a streaming response
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiMessageInput"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}
...
  /ghostwatch/picks:
    get:
      operationId: listPicks
      tags: [ghostwatch]
      summary: Get recommended player prop picks
      parameters:
        - name: sport
          in: query
          schema: { type: string }
        - name: riskTier
          in: query
          schema: { type: string, enum: [Safe, Balanced, Aggressive] }
        - name: minConfidence
          in: query
          schema: { type: number }
      responses:
        "200":
          description: List of picks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pick"
...
  /ghostwatch/games:
    get:
      operationId: listGames
      tags: [ghostwatch]
      summary: Today's slate of games
      parameters:
        - name: sport
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of today's games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
...
  /ghostspere/tickets:
    get:
      operationId: listTickets
      tags: [ghostspere]
      summary: Auto-built prop combo tickets from Ghostwatch/Ghost Express picks
      parameters:
        - name: riskTier
          in: query
          schema: { type: string, enum: [Safe, Balanced, Aggressive] }
      responses:
        "200":
          description: Tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
  /ghostspere/send-email:
    post:
      operationId: sendTicketEmail
      tags: [ghostspere]
      summary: Send selected tickets to user email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailRequest"
      responses:
        "200":
          description: Email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailResult"

  # ─── Ghostobservation Social & News Intel ─────────────────────────────────
...
components:
...
  schemas:
...
    OpenaiConversation:
      type: object
      required: [id, title, createdAt]
      properties:
        id: { type: number }
        title: { type: string }
        createdAt: { type: string, format: date-time }
    OpenaiMessage:
      type: object
      required: [id, conversationId, role, content, createdAt]
      properties:
        id: { type: number }
        conversationId: { type: number }
        role: { type: string }
        content: { type: string }
        createdAt: { type: string, format: date-time }
    OpenaiConversationInput:
      type: object
      required: [title]
      properties:
        title: { type: string }
    OpenaiMessageInput:
      type: object
      required: [content]
      properties:
        content: { type: string }
    OpenaiConversationWithMessages:
      type: object
      required: [id, title, createdAt, messages]
      properties:
        id: { type: number }
        title: { type: string }
        createdAt: { type: string, format: date-time }
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
    OpenaiError:
      type: object
      required: [error]
      properties:
        error: { type: string }

    # ─── Ghostwatch ──────────────────────────────────────────────────────────
    Pick:
      type: object
      required: [id, playerName, team, opponent, propType, line, projection, confidence, riskTier, explanation, sport, createdAt]
      properties:
        id: { type: string }
        playerName: { type: string }
        team: { type: string }
        opponent: { type: string }
        sport: { type: string }
        propType: { type: string }
        line: { type: number }
        direction: { type: string, enum: [Over, Under] }
        projection: { type: number }
        confidence: { type: number }
        riskTier: { type: string, enum: [Safe, Balanced, Aggressive] }
        explanation: { type: string }
        socialImpact: { type: ["number", "null"] }
        createdAt: { type: string }
...
    Game:
      type: object
      required: [id, homeTeam, awayTeam, sport, scheduledAt, status]
      properties:
        id: { type: string }
        homeTeam: { type: string }
        awayTeam: { type: string }
        sport: { type: string }
        scheduledAt: { type: string }
        status: { type: string, enum: [Scheduled, Live, Final] }
        homeScore: { type: ["number", "null"] }
        awayScore: { type: ["number", "null"] }
        quarter: { type: ["string", "null"] }
        timeRemaining: { type: ["string", "null"] }

    # ─── Ghost Express ───────────────────────────────────────────────────────
...
    Ticket:
      type: object
      required: [id, riskTier, picks, combinedConfidence, sport, createdAt]
      properties:
        id: { type: string }
        riskTier: { type: string, enum: [Safe, Balanced, Aggressive] }
        picks:
          type: array
          items:
            $ref: "#/components/schemas/Pick"
        combinedConfidence: { type: number }
        sport: { type: string }
        createdAt: { type: string }
    EmailRequest:
      type: object
      required: [email, ticketIds]
      properties:
        email: { type: string }
        ticketIds:
          type: array
          items: { type: string }
        sport: { type: string }
    EmailResult:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean }
        message: { type: string }

    # ─── Ghostobservation ────────────────────────────────────────────────────
...